Rearange test enviroment
